pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "layaos"
version = "0.1.0"
description = "A brain-inspired Cognitive Operating System (LCO) that orchestrates AI modules in rhythm using the LayaCore engine."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
  { name = "Pisit Tiewphopum" },
  { name = "ChatGPT (OpenAI)", email = "noreply@example.com" }
]
keywords = ["ai", "cognitive os", "orchestration", "rhythm", "event-driven", "agent"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: Other/Proprietary License",
  "Operating System :: OS Independent",
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Topic :: Scientific/Engineering :: Artificial Intelligence"
]

dependencies = [
  "numpy>=1.26",
  "scipy>=1.12,<1.13",
  "pydantic>=2.6",
  "opencv-python>=4.9",
  "pyyaml>=6.0",
  "rich>=13.7",
  "loguru>=0.7",
  "fastapi>=0.111",
  "uvicorn[standard]>=0.30",
  "websockets>=12.0",
  "aiohttp>=3.9"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pytest-cov>=5.0",
  "black>=24.0",
  "ruff>=0.5.0",
  "mypy>=1.10"
]
vision = ["opencv-python>=4.9"]
server = ["fastapi>=0.111", "uvicorn[standard]>=0.30", "websockets>=12.0", "aiohttp>=3.9"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["tests*"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
testpaths = ["tests"]
